Allow auto cleaning of HTML to be turned off. Allow cleaning of HTML to be run manually.

+++ src/Sigil/Misc/HTMLPrettyPrint.cpp
@@ -168,7 +168,7 @@ void HTMLPrettyPrint::setInlineTags(QStringList tags)
 void HTMLPrettyPrint::tokenize()
 {
     QChar c;
-    size_t start = 0;
+    int start = 0;
     bool in_comment = false;
     bool collect_tag = false;
     size_t tag_start = 0;
